Added class roller in statroller
