+++ client/src/Views/StatRoller/StatRoller.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
+import ClassRoller from '../ClassRoller';
 import Home from '../Home'
 import './StatRoller.scss'
 
@@ -28,7 +29,7 @@ class StatRoller extends Component {
         baseAC: "",
         align: "",
         initiative: "",
-        placeholder: "TODO"
+        class: ""
     };
     rollStr = event => {
         let result = Math.floor(Math.random() * 6) + 1 + Math.floor(Math.random() * 6) + 1 + Math.floor(Math.random() * 6) + 1;
@@ -108,6 +109,7 @@ class StatRoller extends Component {
         this.rollWis();
         this.rollCha();
         this.rollAlign();
+        this.rollThatClass();
     };
     buttonRoll = () => {
         this.rollStr();
@@ -117,6 +119,7 @@ class StatRoller extends Component {
         this.rollWis();
         this.rollCha();
         this.rollAlign();
+        this.rollThatClass();
     };
     determinePRF(level) {
         let prf;
@@ -157,6 +160,13 @@ class StatRoller extends Component {
         }
         this.setState({
             align: align
+        })
+    }
+    rollThatClass = () => {
+        const classArr = ['Berserker Barbarian', 'Totem Warrior Barbarian', 'College of Lore Bard', 'College of Valor Bard', 'Knowledge Cleric', 'Life Cleric', 'Light Cleric', 'Nature Cleric', 'Tempest Cleric', 'Trickery Cleric', 'War Cleric', 'Moon Druid', 'Arctic Druid', 'Coast Druid', 'Desert Druid', 'Forest Druid', 'Grassland Druid', 'Mountain Druid', 'Swamp Druid', 'Underdark Druid', 'Champion Fighter', 'Battle Master Fighter', 'Eldritch Knight Fighter', 'Open Palm Monk', 'Shadow Monk', '4 Elements Monk', 'Devotion Oath Paladin', 'Ancient Oath Paladin', 'Vengeance Oath Paladin', 'Hunter Ranger', 'Beast Master Ranger', 'Thief Rogue', 'Assassin Rogue', 'Arcane Trickster Rogue', 'Wild Sorcerer', 'Black Dragon Sorcerer', 'Blue Dragon Sorcerer', 'Brass Dragon Sorcerer', 'Bronze Dragon Sorcerer', 'Copper Dragon Sorcerer', 'Gold Dragon Sorcerer', 'Green Dragon Sorcerer', 'Red Dragon Sorcerer', 'Silver Dragon Sorcerer', 'White Dragon Sorcerer', 'Fey Blade-Pact Warlock', 'Fey Chain-Pact Warlock', 'Fey Tome-Pact', 'Old One Blade-Pact Warlock', 'Old One Chain-Pact Warlock', 'Old Tome-Pact Warlock', 'Infernal Blade-Pact Warlock', 'Infernal Chain-Pact Warlock', 'Infernal Tome-Pact Warlock', 'Abjuration Wizard', 'Conjuration Wizard', 'Divination Wizard', 'Enchantment Wizard', 'Evocation Wizard', 'Illusion Wizard', 'Necromancy Wizard', 'Transmutation Wizard'];
+        let rolledClass = classArr[Math.floor(Math.random() * classArr.length)];
+        this.setState({
+            class: rolledClass
         })
     }
     render() {
@@ -217,6 +227,7 @@ class StatRoller extends Component {
                     initiative={this.state.initiative}
                     prf={this.state.prf}
                     align={this.state.align}
+                    rolledClass={this.state.class}
                 />
             </div>
         )
